Avoid comparison of signed with unsigned integer

+++ src/region_neighbor_list.cpp
@@ -78,7 +78,7 @@ bool RegionNeighborList::hasOverlap(double * x, double radius) const {
  */
 void RegionNeighborList::insert(double * x, double radius) {
   int ibin = coord2bin(x);
-  assert(ibin >= 0 && ibin <= bins.size());
+  assert(ibin >= 0 && static_cast<size_t>(ibin) <= bins.size());
 
   bins[ibin].push_back(Particle(x, radius));
   ++ncount;
